Move Duration formatting into dedicated file
